backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
sqlalchemy>=2.0
alembic>=1.14
pydantic>=2.10
yandex-ai-studio-sdk
openpyxl>=3.1
pdfplumber>=0.11
mineru
python-docx>=1.1
python-dotenv>=1.0
httpx>=0.28
loguru>=0.7
pytest
python-multipart>=0.0.18
aiofiles>=24.1
